% Generated by roxygen2 (4.0.0): do not edit by hand
\name{emory_blood_transcript_modules}
\alias{emory_blood_transcript_modules}
\title{Blood transcriptome modules}
\format{A \code{list} of \code{character} of length 346, with one element per
module}
\usage{
data(emory_blood_transcript_modules)
}
\description{
Blood transcriptome modules
}
\references{
http://www.immuneprofiling.org/meni/meni-paper/btm-landing.gsp
 http://www.nature.com/ni/journal/v15/n2/full/ni.2789.html
}

